+++ src/router/index.ts
@@ -1,2 +1,2 @@
 // eslint-disable-next-line import/prefer-default-export
-export { default as TBUser } from './TBUser';
+export { default as TBUser } from './TBUser.routes';
